Move functions to reconciling file
