+++ internal/controller/mission/provider.go
@@ -67,19 +67,6 @@ func (r *MissionReconciler) ConfirmProviderInstalled(ctx context.Context, missio
 	return nil
 }
 
-func UpdatePackageStatus(mission *missionv1alpha1.Mission, provider *cpv1.Provider) {
-	if mission.Status.PackageStatus == nil {
-		mission.Status.PackageStatus = map[string]missionv1alpha1.MissionPackageStatus{}
-	}
-	ps := mission.Status.PackageStatus[provider.Name]
-	for _, c := range provider.Status.Conditions {
-		if c.Type == "Installed" {
-			ps.Installed = string(c.Status)
-		}
-	}
-	mission.Status.PackageStatus[provider.Name] = ps
-}
-
 func (r *MissionReconciler) GetProvider(ctx context.Context, providerName string) (*cpv1.Provider, error) {
 	p := &cpv1.Provider{}
 	err := r.Get(ctx, types.NamespacedName{Name: providerName}, p)
